add better zooming
